requirements.txt
django
djangorestframework
django-extensions
django-stubs
drf-yasg[validation]
psycopg2
pygraphviz
python-dotenv
